Refactor code structure and improve type safety; update build script to skip linting, fix import paths, and enhance email verification component.
