+++ app/(pages)/(content)/layout.tsx
@@ -1,9 +1,10 @@
 import Navbar from "@components/navbar";
+import React from "react";
 
-export default function ContentLayout({children}){
-    return(
+export default function ContentLayout({ children }: { children: React.ReactNode }) {
+    return (
         <div className="flex flex-col">
-            <Navbar/>
+            <Navbar />
             {children}
         </div>
     )
