+++ app/layout.tsx
@@ -2,7 +2,7 @@ import { Metadata } from "next";
 import { Roboto } from "next/font/google";
 import "@styles/globals.css";
 import { AppProvider } from "@contexts/appContext";
-import { AuthProvider } from "@contexts/AuthContext";
+import { AuthProvider } from "@contexts/authContext";
 const roboto = Roboto({ subsets: ["latin"], weight: ["300", "400", "500", "700"] });
 
 export const metadata: Metadata = {
